Refactor user resolvers
* better naming for the two attributes for terms and conditios
* merge `user_management` and `users` to avoid confusion
* remove unnecessary code in userMiddleware
