+++ backend/src/schema/types/type/User.gql
@@ -24,7 +24,7 @@ type User {
   createdAt: String
   updatedAt: String
 
-  hasAgreedToLatestTermsAndConditions: Boolean 
+  termsAndConditionsAgreedVersion: String!
 
   notifications(read: Boolean): [Notification]! @relation(name: "NOTIFIED", direction: "IN")
 
@@ -154,6 +154,7 @@ type Query {
   ): [User]
 
   blockedUsers: [User]
+  currentUser: User
 }
 
 type Mutation {
